test: adapt test to new components
