+++ frontend/src/tests/lib/components/proposal-detail/ProposalProposerActionsEntry.spec.ts
@@ -4,14 +4,14 @@
 
 import type { Proposal } from "@dfinity/nns";
 import { render } from "@testing-library/svelte";
+import ProposalProposerActionsEntry from "../../../../lib/components/proposal-detail/ProposalProposerActionsEntry.svelte";
 import { proposalFirstActionKey } from "../../../../lib/utils/proposals.utils";
 import {
   mockProposalInfo,
   proposalActionMotion,
+  proposalActionNnsFunction21,
   proposalActionRewardNodeProvider,
 } from "../../../mocks/proposal.mock";
-import ProposalProposerActionsEntry
-  from "../../../../lib/components/proposal-detail/ProposalProposerActionsEntry.svelte";
 
 const proposalWithMotionAction = {
   ...mockProposalInfo.proposal,
@@ -28,7 +28,6 @@ describe("ProposalProposerActionsEntry", () => {
     const { getByText } = render(ProposalProposerActionsEntry, {
       props: {
         proposal: proposalWithMotionAction,
-        proposalId: mockProposalInfo.id,
       },
     });
 
@@ -40,7 +39,6 @@ describe("ProposalProposerActionsEntry", () => {
     const { getByText } = render(ProposalProposerActionsEntry, {
       props: {
         proposal: proposalWithMotionAction,
-        proposalId: mockProposalInfo.id,
       },
     });
 
@@ -55,7 +53,6 @@ describe("ProposalProposerActionsEntry", () => {
     const nodeProviderActions = render(ProposalProposerActionsEntry, {
       props: {
         proposal: proposalWithRewardNodeProviderAction,
-        proposalId: mockProposalInfo.id,
       },
     });
 
@@ -66,10 +63,33 @@ describe("ProposalProposerActionsEntry", () => {
     const motionActions = render(ProposalProposerActionsEntry, {
       props: {
         proposal: proposalWithMotionAction,
-        proposalId: mockProposalInfo.id,
       },
     });
 
     expect(motionActions.queryAllByTestId("json").length).toBe(0);
+  });
+
+  it("should render nnsFunction id", () => {
+    const proposalWithNnsFunctionAction = {
+      ...mockProposalInfo.proposal,
+      action: proposalActionNnsFunction21,
+    } as Proposal;
+
+    const { getByText } = render(ProposalProposerActionsEntry, {
+      props: {
+        proposal: proposalWithNnsFunctionAction,
+      },
+    });
+
+    const [key, value] = Object.entries(
+      (
+        proposalWithNnsFunctionAction?.action as {
+          ExecuteNnsFunction: object;
+        }
+      ).ExecuteNnsFunction
+    )[0];
+
+    expect(getByText(key)).toBeInTheDocument();
+    expect(getByText(value)).toBeInTheDocument();
   });
 });
