+++ frontend/src/tests/lib/components/proposal-detail/ProposalProposerPayloadEntry.spec.ts
@@ -6,10 +6,8 @@ import type { Proposal } from "@dfinity/nns";
 import { render, waitFor } from "@testing-library/svelte";
 import { mock } from "jest-mock-extended";
 import { NNSDappCanister } from "../../../../lib/canisters/nns-dapp/nns-dapp.canister";
-import ProposalProposerActionsEntry from "../../../../lib/components/proposal-detail/ProposalProposerActionsEntry.svelte";
+import ProposalProposerPayloadEntry from "../../../../lib/components/proposal-detail/ProposalProposerPayloadEntry.svelte";
 import { proposalPayloadsStore } from "../../../../lib/stores/proposals.store";
-import { getNnsFunctionKey } from "../../../../lib/utils/proposals.utils";
-import en from "../../../mocks/i18n.mock";
 import {
   mockProposalInfo,
   proposalActionNnsFunction21,
@@ -20,7 +18,7 @@ const proposalWithNnsFunctionAction = {
   action: proposalActionNnsFunction21,
 } as Proposal;
 
-describe("ProposalProposerActionsEntry", () => {
+describe("ProposalProposerPayloadEntry", () => {
   const nnsDappMock = mock<NNSDappCanister>();
   nnsDappMock.getProposalPayload.mockResolvedValue({});
   jest.spyOn(NNSDappCanister, "create").mockImplementation(() => nnsDappMock);
@@ -29,46 +27,15 @@ describe("ProposalProposerActionsEntry", () => {
 
   afterAll(jest.clearAllMocks);
 
-  it("should render nnsFunction id", () => {
-    const { getByText } = render(ProposalProposerActionsEntry, {
-      props: {
-        proposal: proposalWithNnsFunctionAction
-      },
-    });
-
-    const [key, value] = Object.entries(
-      (
-        proposalWithNnsFunctionAction?.action as {
-          ExecuteNnsFunction: object;
-        }
-      ).ExecuteNnsFunction
-    )[0];
-
-    expect(getByText(key)).toBeInTheDocument();
-    expect(getByText(value)).toBeInTheDocument();
-  });
-
-  it("should render nnsFunction name", () => {
-    const { getByText } = render(ProposalProposerActionsEntry, {
-      props: {
-        proposal: proposalWithNnsFunctionAction
-      },
-    });
-
-    const nnsFunctionKey = getNnsFunctionKey(proposalWithNnsFunctionAction);
-    const fnName = en.nns_functions[nnsFunctionKey as string];
-
-    expect(getByText(fnName)).toBeInTheDocument();
-  });
-
   it("should trigger getProposalPayload", async () => {
     const spyGetProposalPayload = jest
       .spyOn(nnsDappMock, "getProposalPayload")
       .mockImplementation(async () => ({}));
 
-    render(ProposalProposerActionsEntry, {
+    render(ProposalProposerPayloadEntry, {
       props: {
-        proposal: proposalWithNnsFunctionAction
+        proposal: proposalWithNnsFunctionAction,
+        proposalId: mockProposalInfo.id
       },
     });
